Changes width of mikid.
This width complies with the expectations of mikmeta.

+++ src/include/miknet/mikid.h
@@ -7,6 +7,6 @@
  *  Returns an identifier. This identifier is guarunteed not to have been given
  *  to any preceding caller. No future callers will receive it either.
  */
-uint64_t mikid();
+uint16_t mikid();
 
 #endif  /* MIKNET_MIKID_H_ */
